pyproject.toml
[project]
name = "review-radar"
version = "0.1.0"
description = "App 用户评论洞察 Agent — 输入 App 名字，自动抓取评论、多维度分析、输出洞察报告"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0,<3.0.0",
    "rich>=13.0.0,<15.0.0",
    "google-play-scraper>=1.2.7,<2.0.0",
    "httpx>=0.27.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
web = [
    "streamlit>=1.40.0,<2.0.0",
    "plotly>=5.24.0,<6.0.0",
    "google-cloud-storage>=2.18.0,<3.0.0",
]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "ruff>=0.4.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["review_radar*"]

[project.scripts]
review-radar = "review_radar.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
